Move map stuff into home panel ui fn
